+++ src/main/java/step2/domain/LottoResult.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import java.util.HashMap;
 import java.util.Map;
