+++ src/test/java/step2/controller/LottoMachineTest.java
@@ -1,7 +1,8 @@
-package step2;
+package step2.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.controller.LottoMachine;
 
 import static org.assertj.core.api.Assertions.assertThat;
