+++ src/test/java/step2/domain/LottoGeneratorTest.java
@@ -1,7 +1,8 @@
-package step2;
+package step2.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.LottoGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
