+++ src/test/java/step2/domain/LottoNumberTest.java
@@ -1,10 +1,11 @@
-package step2;
+package step2.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.LottoNumber;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static step2.LottoNumber.LOTTO_NUMBER_RANGE_EXCEPTION_MESSAGE;
+import static step2.domain.LottoNumber.LOTTO_NUMBER_RANGE_EXCEPTION_MESSAGE;
 
 public class LottoNumberTest {
