+++ src/test/java/step2/domain/LottoTest.java
@@ -1,10 +1,13 @@
-package step2;
+package step2.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import step2.domain.Lotto;
+import step2.domain.LottoGenerator;
+import step2.domain.LottoNumber;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -13,8 +16,8 @@ import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static step2.Lotto.LOTTO_NUMBERS_COUNT_EXCEPTION_MESSAGE;
-import static step2.Lotto.LOTTO_NUMBERS_DUPLICATE_EXCEPTION_MESSAGE;
+import static step2.domain.Lotto.LOTTO_NUMBERS_COUNT_EXCEPTION_MESSAGE;
+import static step2.domain.Lotto.LOTTO_NUMBERS_DUPLICATE_EXCEPTION_MESSAGE;
 
 public class LottoTest {
