+++ src/test/java/step2/domain/LottoTierTest.java
@@ -1,7 +1,8 @@
-package step2;
+package step2.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.LottoTier;
 
 import static org.assertj.core.api.Assertions.assertThat;
